+++ server/controllers/web/categories.js
@@ -39,11 +39,27 @@ exports.addCategory = async (req, res) => {
 
 // delete category
 exports.deleteCategory = async (req, res) => {
+  // Lấy thông tin của nhân viên từ JWT token
+  const token = req.headers.authorization.split(" ")[1];
+  const decodedToken = jwt.verify(token, process.env.JWT_SECRET);
+  const loggedInEmployeeId = decodedToken.userId;
+
+  const categoryId = req.params.id;
+
   try {
-    const category = await Category.findById(req.params.id);
+    // Tìm kiếm thông tin của nhân viên dựa trên id
+    const loggedInEmployee = await Employee.findById(loggedInEmployeeId);
+
+    if (!loggedInEmployee) {
+      // Nếu không tìm thấy nhân viên, trả về lỗi
+      return res.status(404).json({ message: "Employee not found" });
+    }
+
+    const category = await Category.findById(categoryId);
     if (!category) {
       return res.status(404).json({ message: "Category not found" });
     }
+
     await category.deleteOne();
     res.json({ message: "Category deleted successfully" });
   } catch (error) {
@@ -53,8 +69,23 @@ exports.deleteCategory = async (req, res) => {
 
 // update category
 exports.updateCategory = async (req, res) => {
+  // Lấy thông tin của nhân viên từ JWT token
+  const token = req.headers.authorization.split(" ")[1];
+  const decodedToken = jwt.verify(token, process.env.JWT_SECRET);
+  const loggedInEmployeeId = decodedToken.userId;
+
+  const categoryId = req.params.id;
+
   try {
-    const category = await Category.findById(req.params.id);
+    // Tìm kiếm thông tin của nhân viên dựa trên id
+    const loggedInEmployee = await Employee.findById(loggedInEmployeeId);
+
+    if (!loggedInEmployee) {
+      // Nếu không tìm thấy nhân viên, trả về lỗi
+      return res.status(404).json({ message: "Employee not found" });
+    }
+
+    const category = await Category.findById(categoryId);
     if (!category) {
       return res.status(404).json({ message: "Category not found" });
     }
